producer/requirements.txt
requests
beautifulsoup4
lxml
kafka-python
schedule
python-dotenv